Adding AdapterFinderBean and interceptor functionality.
AdapterFinderBean and AdapterFinderInterceptor allow for similar
functionality to ServiceLocatorFactoryBean but without the call from
the client to find the appropriate service bean for each call.

It would be anticipated that spring-boot functionality would be
created in the future to search the bean definition registry after
the registry is created to find finder beans and build proxies for
the interfaces exposed by them and register those implementations
within the application context.

Signed-off-by: Joe Chambers <25091819+joe-chambers@users.noreply.github.com>